Fix bug in codegen for user defined seq
Exporting a sequence constructor did not work because BUILDS was not removed from code.  Add RECORDS which moves the last arg to the first and works like RECORD otherwise.  Change BUILDS to BUILDSEQ which has the  sequence type which allows the BUILDSEQ to be changed to a RECORDS when building abstract types.